chore!: move types from `Candoumbe.Types.Numerics` into a dedicated `Candoumbe.Types.Numerics` project.
