+++ src/Candoumbe.Types.Numerics/NonNegativeInteger.cs
