+++ test/Candoumbe.Types.Numerics.UnitTests/Candoumbe.Types.Numerics.UnitTests.csproj
@@ -1,12 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk">
-  
+
   <Import Project="..\..\tests.props" />
   <PropertyGroup>
-    <TargetFrameworks>net8.0</TargetFrameworks>
+    <TargetFrameworks>net8.0;</TargetFrameworks>
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Candoumbe.Types\Candoumbe.Types.csproj" />
+    <ProjectReference Include="..\..\src\Candoumbe.Types.Numerics\Candoumbe.Types.Numerics.csproj" />
   </ItemGroup>
 
   <ItemGroup>
@@ -15,7 +15,5 @@
     </None>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Numerics\" />
-  </ItemGroup>
+
 </Project>
