+++ test/Candoumbe.Types.Numerics.UnitTests/xunit.runner.json
@@ -1,0 +1,4 @@
+﻿{
+  "methodDisplay": "method",
+  "methodDisplayOptions": "all"
+}
