+++ test/Candoumbe.Types.UnitTests/Generators/ValueGenerators.cs
@@ -2,7 +2,6 @@
 // Licenced under GNU General Public Licence, version 3.0"
 
 using Candoumbe.Types.Calendar;
-using Candoumbe.Types.Numerics;
 
 using FsCheck;
 using FsCheck.Fluent;
@@ -144,14 +143,4 @@ internal static class ValueGenerators
     public static Arbitrary<Array> Arrays() => ArbMap.Default.ArbFor<int>().Generator.ArrayOf()
                                                   .Select(Array (numbers) => numbers)
                                                   .ToArbitrary();
-
-    public static Arbitrary<PositiveInteger> PositiveIntegers() => ArbMap.Default.ArbFor<PositiveInt>()
-                                                                         .Generator
-                                                                         .Select(value => PositiveInteger.From(value.Item))
-        .ToArbitrary();
-
-    public static Arbitrary<NonNegativeInteger> NonNegativeIntegers() => ArbMap.Default.ArbFor<NonNegativeInt>()
-                                                                         .Generator
-                                                                         .Select(value => NonNegativeInteger.From(value.Item))
-        .ToArbitrary();
 }
